Refactor: share buffer reader between the mono and stereo sample seq nodes
